ToDo-App/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-core==1.5.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.3.0
azure-mgmt-subscription==3.1.1
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.46
botocore==1.38.46
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.14
fastapi-cli==0.0.7
Flask==3.1.1
flask-babel==4.0.0
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsonformatter==0.3.4
keyring==25.6.0
ldap3==2.9.1
libgravatar==1.0.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
packaging==25.0
paramiko==3.5.1
passlib==1.7.4
pgadmin4==9.4
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
s3transfer==0.13.0
SecretStorage==3.3.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sshtunnel==0.4.0
starlette==0.46.2
tomli==2.2.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.4.0
user-agents==2.2.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
zipp==3.23.0
zstandard==0.23.0
